test(codex): isolate authenticated hook coverage
